+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ Changelog
 Forthcoming
 -----------
 * [tests] moved from unittest to pytest
-* ...
+* [launch] dashing/eloquent handling of emulate_tty
 
 1.0.5 (2019-10-04)
 ------------------